feat: Game State moved into own script, contract not in DDOL
